+++ scripts/hooks/session-end.cjs
@@ -19,7 +19,7 @@ const {
   writeFile,
   replaceInFile,
   log
-} = require('../lib/utils');
+} = require('../lib/utils.cjs');
 
 async function main() {
   const sessionsDir = getSessionsDir();